+++ src/components/matriculas/cuenta/ResumenCuenta.vue
@@ -3,8 +3,16 @@
       <dialog-detalle :item="item_selected" ref="show_detalle">
       </dialog-detalle>
 
-      <br>
-      <v-layout row wrap>
+      <v-layout row class="mt-4">
+        <v-flex xs12>
+          <v-btn class="green right" dark @click="showAddBoleta">
+            <v-icon class="mr-2">add</v-icon>
+            Nueva Boleta
+          </v-btn>
+        </v-flex>
+      </v-layout>
+
+      <v-layout row wrap class="mt-4">
         <v-flex xs1 class="mt-5 mx-5">
           <b>Filtrar:</b>
         </v-flex>
@@ -25,20 +33,15 @@
               no-data-text="No hay datos"
               no-results-text="No hay datos"
               :rows-per-page-items="[25,30,35]"
+              v-bind:pagination.sync="pagination"
           >
-            <template slot="headers" slot-scope="props">
-              <th v-for="header of props.headers" :key="header.value" class="pa-3 text-xs-left">
-                <b>{{ header.text }}</b>
-              </th>
-            </template>
             <template slot="items" slot-scope="props">
               <tr>
                 <td>{{ props.item.fecha | fecha }}</td>
                 <td>{{ props.item.fecha_vencimiento | fecha }}</td>
-                <td v-if="props.item.tipo == 'boleta'">{{ props.item.tipo_comprobante.descripcion }}</td>
-                <td v-if="props.item.tipo == 'comprobante'">Recibo</td>
-                <td>{{ props.item.tipo == 'boleta' ? props.item.total : '' }}</td>
-                <td>{{ props.item.tipo == 'comprobante' ? props.item.importe_cancelado : '' }}</td>
+                <td>{{ props.item.descripcion }}</td>
+                <td>{{ props.item.debe }}</td>
+                <td>{{ props.item.haber }}</td>
                 <td>
                   <v-btn fab dark small color="blue" @click="verDetalle(props.item)">
                     <v-icon>launch</v-icon>
@@ -68,28 +71,58 @@
           </v-data-table>
         </v-flex>
       </v-layout>
+
+      <v-dialog v-model="show_addboleta" fullscreen transition="dialog-bottom-transition" :overlay="false">
+        <v-card>
+          <v-toolbar dark class="blue">
+            <v-toolbar-title class="white--text">Nueva Boleta</v-toolbar-title>
+            <v-spacer></v-spacer>
+            <v-btn icon @click="show_addboleta = false">
+              <v-icon>close</v-icon>
+            </v-btn>
+          </v-toolbar>
+          
+          <nueva-boleta
+            :id-matricula="id"
+            @cancelar="show_addboleta = false"
+            @update="nuevaBoleta"
+          ></nueva-boleta>
+
+        </v-card>
+      </v-dialog>      
     </v-card>
 </template>
 
 <script>
-import axios from '@/axios';
-import * as utils from '@/utils';
-import moment from 'moment';
+import axios from '@/axios'
+import * as utils from '@/utils'
+import moment from 'moment'
+import { Header } from '@/model'
 import DatosBasicos from '@/components/matriculas/DatosBasicos';
 import DialogDetalle from '@/components/matriculas/cuenta/detalles/DialogDetalle';
+import NuevaBoleta from '@/components/matriculas/cuenta/NuevaBoleta';
 import InputFecha from '@/components/base/InputFecha';
 import ValidatorMixin from '@/components/mixins/ValidatorMixin';
 
 const headers = [
-  { text: 'Fecha' },
-  { text: 'Fecha de Vencimiento' },
-  { text: 'Descripción' },
-  { text: 'Debe' },
-  { text: 'Haber' }
+  Header('Fecha', 'fecha', true),
+  Header('Fecha de Vencimiento', 'fecha_vencimiento', true),
+  Header('Descripción', 'descripcion', true),
+  Header('Debe', 'debe', true),
+  Header('Haber', 'haber', true),
+  Header('', 'detalle')
 ]
 
-const orderByFecha = function(a, b) {
-  return moment(a.fecha_vencimiento).diff(b.fecha_vencimiento, 'days')
+const sortByFecha = (atributo) => {
+  return (a, b) => moment(a[atributo]).diff(b[atributo], 'days');
+}
+
+const sortByString = (atributo) => {
+  return (a, b) => a[atributo].localeCompare(b[atributo]);
+}
+
+const sortByNumber = (atributo) => {
+  return (a, b) => b[atributo] - a[atributo];
 }
 
 export default {
@@ -100,9 +133,15 @@ export default {
   data () {
     return {
       resumen: [],
+      resumen_original: [],
       boletas: [],
       show_detalle: false,
+      show_addboleta: false,
       item_selected: null,
+      pagination: {
+        sortBy: 'fecha_vencimiento',
+        descending: false
+      },
       filtros: {
         fecha_desde: moment().startOf('year').format("DD/MM/YYYY"),
         fecha_hasta: moment().endOf('year').format("DD/MM/YYYY")
@@ -144,6 +183,15 @@ export default {
           this.updateBoletas();
       },
       deep: true
+    },
+
+    'pagination.sortBy': function(sortBy) {
+      if (sortBy) {
+        if (sortBy.includes('fecha')) this.resumen = this.resumen.sort(sortByFecha(sortBy));
+        else if (sortBy == 'debe' || sortBy == 'haber') this.resumen = this.resumen.sort(sortByNumber(sortBy));
+        else this.resumen = this.resumen.sort(sortByString(sortBy));
+      }
+      else this.resumen = utils.clone(this.resumen_original);
     }
   },
 
@@ -173,12 +221,18 @@ export default {
      .then(([boletas, comprobantes]) => {
        let resumen = boletas.data.map(b => {
          b.tipo = 'boleta';
+         b.debe = b.total;
+         b.descripcion = b.tipo_comprobante.descripcion;
          return b;
        }).concat(comprobantes.data.map(c => {
          c.tipo = 'comprobante';
+         c.haber = c.importe_cancelado;
+         c.descripcion = 'Recibo';
          return c;
        }));
-       this.resumen = resumen.sort(orderByFecha);
+       this.resumen = resumen;
+       this.resumen_original = utils.clone(resumen);
+       this.resumen = this.resumen.sort(sortByFecha('fecha_vencimiento'));
      })
      .catch(e => console.error(e));
     },
@@ -188,12 +242,22 @@ export default {
       this.$refs.show_detalle.mostrar();
     },
 
+    showAddBoleta: function() {
+      this.show_addboleta = true;
+    },
+
+    nuevaBoleta: function() {
+      this.updateBoletas();
+      this.show_addboleta = false;
+    }
+
   },
 
   components: {
     DatosBasicos,
     DialogDetalle,
-    InputFecha
+    InputFecha,
+    NuevaBoleta
   }
 
 }