+++ src/stores/listados/Matriculas.js
@@ -2,7 +2,9 @@ export default {
   state: {
       pagination: {
         page: 1,
-        rowsPerPage: 25
+        rowsPerPage: 25,
+        descending: null,
+        sortBy: null
       },
 
       filtros: {
@@ -25,6 +27,15 @@ export default {
 
   setFiltros(filtros) {
     this.state.filtros = filtros;
+  },
+
+  limpiarFiltros() {
+    this.state.filtros.tipoEntidad = 'profesional';
+    this.state.filtros.estado = 'Habilitado';
+    this.state.filtros.numero = '';
+    this.state.filtros.profesional.dni = '';
+    this.state.filtros.profesional.apellido = '';
+    this.state.filtros.empresa.nombre = '';
   }
 }